microservices_backend/requirements.txt
flask
python-dotenv
requests
flask_cors
tenacity
pandas
sqlalchemy
pymysql
pydantic
fastapi
uvicorn
pymongo
isodate
rich

